feat: 코스리뷰 전체보기 api 설계 및 코스 평점 타입 수정

+++ src/main/java/umc/catchy/domain/courseReview/dto/response/PostCourseReviewResponse.java
@@ -36,7 +36,7 @@ public class PostCourseReviewResponse {
     @AllArgsConstructor
     @NoArgsConstructor
     public static class courseReviewAllResponseDTO{
-        Float courseRating;
+        Double courseRating;
         Integer totalCount;
         CourseReviewSliceResponse courseReviewSliceResponse;
     }
